fix: show full day name (Thursday instead of Thu) and hide Awaiting Review badge from WFH detail modal

+++ src/components/attendance/EmployeeWFHHistoryTable.tsx
@@ -10,7 +10,7 @@ export const formatWFHDate = (dateString: string) => {
     if (!dateString) return "-";
     const date = new Date(dateString);
     return new Intl.DateTimeFormat('en-GB', {
-        weekday: 'short',
+        weekday: 'long',
         day: '2-digit',
         month: 'short',
         year: 'numeric',
